Fixed bounce (further testing needed in game loop)
